refactor(hypr): rename workstation profile to default, gitignore machine files
Machine profiles now use default.lua as baseline/template instead of
workstation.lua, matching multi-machine use (not just one workstation).
config/machines/.gitignore untracks everything but default.lua, init.lua,
README so personal hostname profiles stay local.
