+++ home/hypr/.config/hypr/config/machines/.gitignore
@@ -1,0 +1,5 @@
+*
+!.gitignore
+!init.lua
+!default.lua
+!README.md
