+++ home/hypr/.config/hypr/config/machines/default.lua
